requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
Pillow
scikit-learn
